refactor(app): extract useLoadingState hook (f64)
Moves isLoading, isInitialAnalysisPending, loadingMessage, loadingProgress,
loadCancelRef, the analysis-worker latch effect, and handleCancelLoad out of
App.tsx into src/hooks/useLoadingState.ts.
